Remove Key reference in Gamestatepane
The current header image of the gamestate panel informs the player that he/she should press G to close it (which is then the only interaction the player has with that panel, ever. ) Given that
1) G is relatively valuable keyboard real estate
2) it is just hugely bad practice to lock in keybinds by embedding references in artwork
I removed the reference to the closing hotkey in the linked png file.

accompanied is an mdl file with an ImportImageFromFile command.